[Refactor] #495 - ProfileVisibility ViewModel 삭제
